Further split tsumugu-net crate
