+++ tsumugu-parser/src/parser/fancyindex.rs
@@ -1,9 +1,8 @@
 // Nginx fancyindex parser
 
-use crate::{
-    client::HttpClient,
-    listing::{FileSize, FileType, ListItem},
-};
+use crate::listing::{FileSize, FileType, ListItem};
+
+use tsumugu_net::client::HttpClient;
 
 use super::*;
 use anyhow::Result;
